requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
confection==0.0.4
conllu==4.5.2
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
executing==1.2.0
filelock==3.11.0
frozenlist==1.3.3
fsspec==2023.6.0
gensim==4.3.1
huggingface-hub==0.15.1
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
langcodes==3.3.0
langdetect==1.0.9
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.24.2
openai==0.27.8
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
preshed==3.0.8
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.6.1
pydantic==1.10.7
pyenchant==3.2.2
Pygments==2.15.0
pysbd==0.3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==0.10.1
wcwidth==0.2.6
yarl==1.9.2
